remove unused definitions and cleanup descriptor files
